refactor(web): drop the generated checklist, relocate the verdict overlay
Addresses three points on the assessment tool:
- Stop the save reloading the page: unwatch vpatVerdicts.json so writing a
  verdict no longer trips Vite HMR; the /assess page already updates from
  the POST response.
- Remove the generated manual-assessment.md and its freshness/drift guards;
  the /assess tool and /accessibility page supersede the static checklist.
  manualAssessmentDoc.ts becomes assessmentGuidance.ts (guidance data only).
- Move the mutable overlay out of the pure-logic util leaf to
  accessibility/vpatVerdicts.json, beside the other a11y assets.

Also reject an empty remark on save (and disable the button to match), so a
verdict always carries the note that is its whole point.
